Renaming Stats_Window to Top_App_Bar and creating a stats window component that contains all the stats for the app
